requirements.txt
langgraph>=0.0.25
langchain>=0.1.0
langchain-google_genai
chromadb>=0.4.22
sentence-transformers>=2.2.2
streamlit>=1.25.0
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.26.0
pydantic>=1.10.0
pytest>=7.0.0